requirements.txt
azure-storage-blob==12.19.0
azure-core==1.29.5
python-dotenv==1.0.0
